Check for transfer-encoding: chunked
Envoy (and other proxies that enforce RFC compliance) don't like to see "transfer-encoding: chunked" and "content-length" in the same response.

The "read-stream-all" method is concatenating all the chunks into one response, so propagating the header is wrong.  It's now a single response.

That's not ideal, because now we don't stream and become RAM-dependent.